+++ demisto_sdk/__main__.py
@@ -638,10 +638,6 @@ def secrets(config, **kwargs):
 )
 @click.option("-dt", "--docker-timeout", default=60,
               help="The timeout (in seconds) for requests done by the docker client.", type=int)
-@click.option("-idp", "--id-set-path", help="Path to id_set.json, relevant for when using the "
-                                            "--check-dependent-api-module flag.",
-              type=click.Path(resolve_path=True),
-              default='Tests/id_set.json')
 @click.option("-cdam", "--check-dependent-api-module", is_flag=True,
               help="Run unit tests and lint on all packages that "
               "are dependent on the found "
@@ -663,6 +659,7 @@ def lint(**kwargs):
                   log_path=kwargs.get('log_path'))  # type: ignore[arg-type]
 
     check_configuration_file('lint', kwargs)
+    print(f"kwargs.get('id_set_path') {kwargs.get('id_set_path')}")
     lint_manager = LintManager(
         input=kwargs.get('input'),  # type: ignore[arg-type]
         git=kwargs.get('git'),  # type: ignore[arg-type]
@@ -671,7 +668,6 @@ def lint(**kwargs):
         quiet=kwargs.get('quiet'),  # type: ignore[arg-type]
         prev_ver=kwargs.get('prev_ver'),  # type: ignore[arg-type]
         json_file_path=kwargs.get('json_file'),  # type: ignore[arg-type]
-        id_set_path=kwargs.get('id_set_path'),  # type: ignore[arg-type]
         check_dependent_api_module=kwargs.get('check_dependent_api_module'),  # type: ignore[arg-type]
     )
     return lint_manager.run(