+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 * Fixed an issue where modeling rules with arbitrary whitespace characters were not parsed correctly.
 * Added support for the **nativeImage** key for an integration/script in the **prepare-content** command.
+* Updated **lint** to use graph instead of id_set when running with `--check-dependent-api-module` flag.
 
 ## 1.8.1
 * Fixed an issue where **format** created duplicate configuration parameters.